move post scraping logic to controller file
